refactor(engine): relocate surrogate.py out of engine/ to ml/
External audit finding B1 (P1): src/sisyphus/engine/surrogate.py:129 had
`params.node_enzymes("liver")` — a string match on an organ name inside
engine/. This contradicted the README §Design principles claim: "No string
matching on organ names, enzyme names, or drug names exists in `engine/`."

Mitigating context: surrogate.py is opt-in (`backend="surrogate"`) and not
exported from engine/__init__.py, so the production scipy LSODA path was
unaffected. But the file's location in `engine/` violated the binary
invariant claim.

Resolution: git mv to src/sisyphus/ml/surrogate.py.

The "liver" string match remains in surrogate.py (the surrogate was trained
on the standard reference_man graph, hepatic-only assumption baked into its
input contract) — this is now an internal weakness of ml/surrogate.py, not
an engine/ invariant violation.

Importers updated:
- src/sisyphus/engine/uncertainty.py:218 (lazy import; opt-in surrogate
  backend in MC propagation — soft layer dependency, acceptable)
- src/sisyphus/regimen/tdm_sbi.py:264 (lazy import; opt-in TDM acceleration)
- src/sisyphus/sbi/multi_drug.py:153 (docstring reference)
- tests/unit/test_surrogate_features.py:12
- scripts/{validate_surrogate_production,train_surrogate}.py
- docs/{surrogate_ood_fix,sbi_poc_results,next_steps_plan}.md (path refs)

Verification:
- `grep -rn 'engine.surrogate\|engine/surrogate' src/ tests/ scripts/` → 0 hits
- `grep -rn '"liver"\|"kidney"\|...' src/sisyphus/engine/` → 0 hits (the only
  remaining string organ-references are `observation_node="venous_blood"`
  default arg in solver.py:111, solver_jax.py:290, uncertainty.py:183 —
  caller-hint defaults that the random-string-rename test would override)
- pytest tests/unit/test_surrogate_features.py → 7 passed
- pytest tests/integration tests/regression → 110 passed, 7 xfailed
  (bit-identical pre-move sweep)

No headline AAFE impact (surrogate is opt-in; scipy LSODA path unchanged).
